requirements.txt
google-genai>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.1
faiss-cpu>=1.7.4
pdfminer.six>=20221105
pillow>=9.0.0
gradio>=4.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
tqdm>=4.65.0
pyyaml>=6.0.0
